split babel stuff into its own module
